Major refactor to using migrated DOM nodes.
Because so much of the Foundry functionality relies on DOM node event
listeners and client side hooks, having a completely separate window
loaded was breaking other modules that relied on a client having access
to a sigular DOM. This change modifies the window.open to instead of
loading the entire application froms scratch. Instead of just moves the
applications DOM nodes into the child window. This has it's own
problems, but it's behavior lines up much more with what module authors
are expecting. In addition it uses judicious use of cloning to ensure
that the child document is as close as possible to the parent instead of
having be constructed by hand. I.e. it clones the entire head node.

Ideally this really should be part of foundary core. As then the Hooks
system could be multi-window aware. And opening a new window could know
at a foundry level whether it is part of a logical existing instance.
http://golden-layout.com/ is an excellent template to start from in
terms of modelling the idea of a multi window version of foundry. The
main decision being whether you just alter the hooks/socket emit system
to broadcast across windows, or alter the data persistence model to keep
a copy of Application state in sync via localstorage is the main one.