Fixes for editor and size.
Added a hack to ensure tinyMCE works with our popouts.
We need to make sure tinyMCE is initialized before any popout is
created. This is because it lazy loads it's plugin scripts, which are
then not available in the popout document if the editor button is
clicked. So instead we create a minimal editor and attach+remove it from
the DOM in our initnitialization.

I have also cleaned up the close popup to ensure that the original
window size is preserved. Before this fix it was appearing as a full
screen window the next time the same application was opened.