Fix dialogs moving position on re-render (Fixed: #31)
This change fixes an issue where dialogs (logical sub applications) move
when re-rendering. The solution is to just overwrite the setPosition
function of these dialogs to be a no-op. This is safe to do as the
dialogs can only be closed (if they are in a popped out context) and
not *PopInned* so we will never end up with an application in an
inconsistent state.