Fix broken escape key handling (Fixes #34)
Due to the escape key handler in foundry.js assuming that all windows
in the ui.windows list are visible and closing them before the rest of
the handlers are called (i.e. deselect token or main menu handler). We
have to overload the ownKeys method of the windows.ui proxy object in
order to *hide* popped out windows from the foundry.js handler.