Copy root document styles to popout document (Fixes #38)
Some modules use CSS variables to customize styles.
They can be either locally element scoped or on the root element
`<html>`. If they are locally scoped then they would already work with
PopOut!. However it appears that most tutorials and guides say you
should set the variables at the root of the document. This makes sense
for regular css variables on regular websites, however breaks when
interacting with the PopOut! module because they will not be available
in the new popped out window. This patch copies **all** the root
`<html>` elements styles to the new popped out window as that follows
the principle of least surprise for module authors, even if it is
technically not clear from a css specification/best practices level.