Add compatibility mode for fancy character sheets (Fixes #37)
Some character sheet extensions add fancy decorations or other
non-standard styling. This breaks PopOut! because it makes the
assumption that all html elements will appear within the height and
width of the applications html element. This is not true for example
if absolute positioning is used. In order to handle this use case, a
compatibility mode is introduced in this patch. It works by enumerating
every single html child element of the application and calculating the
total bounding box (think convex hull) of them. It then uses these new
dimensions in place of the application defined ones. However this may
have a serious performance impact on older hardware, or character sheets
with thousands of elements. Therefore it is a togglable and off by
default.