Fix implicit type conversion bug (Fixes #44)
There is a subtle bug in how we reflect the ui.windows object.
The intention with the last fix was to preserve escape behavior from
before where it would allow for opening the escape menu, so we had to
ensure that the only keys returned from windows.ui object were **not**
popped out applications. However the Map we use to track those apps is
has integer keys, where as the Reflect.ownKeys method returns strings.
Thus we need to convert them to integers before filtering otherwise
we end up in the situation where no windows are being returned from
`Object.keys(window.ui.windows)`.