Fix bug with PDFoundry compatibility
There was an issue where windows using PDFoundry if popped out, then
close by clicking the browser close button, would remain in a broken
state and never be able to re-open unless the `Esc` key is pressed.
This is because PDFoundry uses 2 application to represent a single
window and when one of the is closed in the popout, the other remains
open in the main window. The fix is to add a dual of the open handler,
to perform the expected cleanup operation.