Intercept BringToTop (Fixes #58)
Interactions with dragging items onto a character sheet change it's
zIndex to put it in front of any open dialogs. This is confusing
if someone is dragging for example an item into an item container
that is open in a popout as it will now be behind the character sheet
in the popped out window and effectively inaccessible. We solve this
by making sure that the popped out sheet is always behind any dialogs.