Prevent wrong base URL
If the pdf path is only provided as a relative path, pdfjs will add the base from window.location. This can lead to wrong URLs as the base url of the poput frame is used instead of the main vtt application.

This makes it so the base url of the main application is used!