Make app function replacements `async`
Both AppV1 / AppV2 methods for `close`, `maximize`, `minimize`, and `render` are `async`. The temporary replacement functions set when popped out are not async and cause chaining issues internal to Foundry when a Promise is expected to be returned. For instance if `render` is called on a popped out window Foundry internally to render will chain on `maximize` with `this.maximize().then(() => this.bringToTop());`. Since the popout replacement functions are not `async` this will fail at the `.then` continuation internal to Foundry and cause an error. 

This PR just makes the appropriate popout replacement functions `async`.